Refactor fetch types to stay compatible with openapi fetch API
