+++ libs/clients/middlewares/src/index.ts
@@ -1,3 +1,4 @@
 export * from './lib/createEnhancedFetch'
+export * from './lib/types'
 export { FetchError } from './lib/FetchError'
 export * from './lib/buildCacheControl'
